moved NSString+SwiftyAttributes
